Updated restart script to check for new commits before running

+++ restart.sh
@@ -1,19 +1,35 @@
 #!/bin/bash
 
-# Stop the FastAPI service
-sudo systemctl stop wealth-sync.service
-
-# Fetch new code from your repository (replace with the appropriate git command)
+# Fetch changes from the remote repository
 git fetch origin
-git reset --hard origin/master
 
+# Check if there are new commits in the remote repository
+LOCAL=$(git rev-parse @)
+REMOTE=$(git rev-parse @{u})
+BASE=$(git merge-base @ @{u})
 
-source myenv/bin/activate
-pip install -r requirements.txt
+if [ $LOCAL = $REMOTE ]; then
+    echo "Up-to-date"
+elif [ $LOCAL = $BASE ]; then
+    echo "Need to pull"
 
-chmod +x main.sh
-chmod +x restart.sh
-chmod +x livefeed.sh
+    # Stop the FastAPI service
+    sudo systemctl stop wealth-sync.service
 
-# Restart the FastAPI service
-sudo systemctl start wealth-sync.service
+    # Reset local code to match remote repository
+    git reset --hard origin/master
+
+    # Activate virtual environment and install requirements
+    source myenv/bin/activate
+    pip install -r requirements.txt
+
+    # Set execute permissions on scripts
+    chmod +x main.sh
+    chmod +x restart.sh
+    chmod +x livefeed.sh
+
+    # Restart the FastAPI service
+    sudo systemctl start wealth-sync.service
+else
+    echo "Diverged"
+fi
